Switch to using YARP as the parser (#41)
* Switch to using YARP as the parser

This parser is going to be backend for Syntax Tree soon anyway, so
we may as well start using it. It comes with much increased speed
and much reduced memory.

* Fix undefined variable in `Analyzer#analyze`

* Introduce `Class/ModuleDefiniton#qualified_namespace` for matching namespaces

Now, that `Class/ModuleDefiniton#namespace` is an array of constants we need another way of comparing and matching namespaces. That's why this commit introduces `Class/ModuleDefiniton#qualified_namespace` so we can compare and match the namespaces as strings.

---------

Co-authored-by: Marco Roth <marco.roth@intergga.ch>